# Common Implemenatation for MatMul and MatMulTran for both aligned and unaligned arrays