--- a/src/Microsoft.ML.CpuMath/CpuAligenedMathUtils.cs
+++ b/src/Microsoft.ML.CpuMath/CpuAligenedMathUtils.cs
@@ -75,40 +75,32 @@ namespace Microsoft.ML.Runtime.Internal.CpuMath
 
         /// <summary>
         /// Matrix multiplication:
-        /// if (add)
-        ///     dst = mat * src
-        /// else
-        ///     dest += mat * src
+        /// dst = mat * src
         /// </summary>
-        /// <param name="add">The addition flag</param>
         /// <param name="mat">The multiplier matrix</param>
         /// <param name="src">The source vector</param>
         /// <param name="dst">The destination vector</param>
-        public static void MatTimesSrc(bool add, ICpuFullMatrix mat, ICpuVector src, ICpuVector dst)
+        public static void MatTimesSrc(ICpuFullMatrix mat, ICpuVector src, ICpuVector dst)
         {
             bool colMajor = typeof(TMatrix) == typeof(CpuAlignedMatrixCol);
             AssertCompatible(mat, src, dst);
             var m = A(mat);
-            CpuMathUtils.MatTimesSrc(colMajor, add, m.Items, A(src).Items, A(dst).Items, m.RunCnt);
+            CpuMathUtils.MatTimesSrc(colMajor, m.Items, A(src).Items, A(dst).Items, m.RunCnt);
         }
 
         /// <summary>
         /// Matrix transpose multiplication:
-        /// if (add)
-        ///     dst = mat' * src
-        /// else
-        ///     dest += mat' * src
+        /// dst = mat' * src
         /// </summary>
-        /// <param name="add">The addition flag</param>
         /// <param name="mat">The multiplier matrix</param>
         /// <param name="src">The source vector</param>
         /// <param name="dst">The destination vector</param>
-        public static void MatTranTimesSrc(bool add, ICpuFullMatrix mat, ICpuVector src, ICpuVector dst)
+        public static void MatTranTimesSrc(ICpuFullMatrix mat, ICpuVector src, ICpuVector dst)
         {
             bool colMajor = typeof(TMatrix) == typeof(CpuAlignedMatrixCol);
             AssertCompatible(mat, dst, src);
             var m = A(mat);
-            CpuMathUtils.MatTimesSrc(!colMajor, add, m.Items, A(src).Items, A(dst).Items, m.RunCnt);
+            CpuMathUtils.MatTimesSrc(!colMajor, m.Items, A(src).Items, A(dst).Items, m.RunCnt);
         }
     }