… src/Microsoft.ML.CpuMath/CpuMathUtils.netcoreapp.cs
… src/Microsoft.ML.CpuMath/CpuMathUtils.netstandard.cs
… src/Microsoft.ML.CpuMath/Sse.cs
… src/Microsoft.ML.CpuMath/SseIntrinsics.cs
… src/Microsoft.ML.CpuMath/Thunk.cs
… src/Microsoft.ML.TimeSeries/AdaptiveSingularSpectrumSequenceModeler.cs
… src/Microsoft.ML.Transforms/RffTransform.cs
… src/Native/CpuMathNative/Avx.cpp
… src/Native/CpuMathNative/Sse.cpp
… test/Microsoft.ML.CpuMath.PerformanceTests/NativePerformanceTests.cs
… test/Microsoft.ML.CpuMath.PerformanceTests/SsePerformanceTests.cs
… test/Microsoft.ML.CpuMath.UnitTests.netcoreapp/UnitTests.cs