Fixes https://github.com/dotnet/machinelearning/issues/1245
For inputs that are not naturally aligned (the alignment is not a multiple of 4), it does exclusively unaligned loads
For all other inputs, it will do at most two unaligned loads (one each for any leading/trailing unaligned elements) and all other loads will be aligned.

cc @tannergooding @eerhardt @danmosemsft @TomFinley