@Anipik 2018-10-10
``` ini
Before
```
|      Method |     Avx  |     Sse   |    Native |
|------------ |---------:|----------:|----------:|
| MatMulX     | 94.02 us | 104.21 us | 95.90  us |
| MatMulTranX | 94.51 us | 110.10 us | 98.10  us |


``` ini
After
```
|      Method |     Avx  |     Sse   |    Native |
|------------ |---------:|----------:|----------:|
| MatMulX     | 94.67 us | 105.62 us | 95.90  us |
| MatMulTranX | 97.25 us | 117.20 us | 98.15 us |




### Observations

- Native implementations are same in performance even with our 2 unaligned loads

- Matmul performance is almost the same but there is about 3-5 % decline in performance in MatmulTrans