@Anipik 2018-10-11
@tannergooding are the performance tests fine ?